% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{cleanCoordinates}
\alias{cleanCoordinates}
\title{Clean coordinates}
\usage{
cleanCoordinates(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "scientificname"
)
}
\arguments{
\item{x}{data}

\item{lon}{longitude}

\item{lat}{latitude}

\item{species}{species name}
}
\value{

}
\description{
Clean coordinates
}
