% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_mod.R
\name{code_mod_server}
\alias{code_mod_server}
\title{Code server}
\usage{
code_mod_server(id, inputArg, orgVars, visDat, remVars, layers, summary)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{input}{Reactive value with the layers to show}

\item{ractiveValues}{A reactive value with primary biodiversity data}
}
\value{

}
\description{
Code server
}
