% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_mod.R
\name{code_mod_ui}
\alias{code_mod_ui}
\title{code UI}
\usage{
code_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{
UI for verbatim code
}
\description{
code UI
}
