% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui.R
\name{comMatrixUI}
\alias{comMatrixUI}
\title{Community matrix UI}
\usage{
comMatrixUI(session)
}
\arguments{
\item{session}{The server session}
}
\value{
community matrix UI
}
\description{
Community matrix UI
}
