% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDef.R
\name{expDef_mod_server}
\alias{expDef_mod_server}
\title{Export definitions server}
\usage{
expDef_mod_server(id, summary, exportData)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{summary}{A reactive value with the summary data}

\item{exportData}{A reacive value with data to export}
}
\value{

}
\description{
Export definitions server
}
