% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDef.R
\name{expDef_mod_ui}
\alias{expDef_mod_ui}
\title{Export definitions UI}
\usage{
expDef_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{
export definitions ui
}
\description{
Export definitions UI
}
