% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expParam_mod.R
\name{expParam_mod_server}
\alias{expParam_mod_server}
\title{Export parameters server}
\usage{
expParam_mod_server(id, summary)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{summary}{A reactive value with the summary data}
}
\value{
export params outputs
}
\description{
Export parameters server
}
