% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expParam_mod.R
\name{expParam_mod_ui}
\alias{expParam_mod_ui}
\title{Export parameters UI}
\usage{
expParam_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{

}
\description{
Export parameters UI
}
