% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resView.R
\name{getColor}
\alias{getColor}
\title{Coloring function}
\usage{
getColor(data, max)
}
\arguments{
\item{data}{A numeric vector}

\item{max}{The maximum value of the vector}
}
\description{
Coloring function
}
