% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDef.R
\name{getExportBirds}
\alias{getExportBirds}
\title{Get export BIRDS}
\usage{
getExportBirds(sb, dim, tr, var, mtd)
}
\arguments{
\item{sb}{a SummarizedBirds-object}

\item{dim}{dimension}

\item{tr}{timeRes}

\item{var}{variable}

\item{mtd}{method}
}
\value{
exported variables from BIRDS::exportBirds
}
\description{
Get export BIRDS
}
