% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mod.R
\name{getGridFromSettings}
\alias{getGridFromSettings}
\title{Get grid from settings}
\usage{
getGridFromSettings(area, gridsize, type, buffer)
}
\arguments{
\item{area}{A spatial polygon}

\item{gridsize}{The size of the grid in km}

\item{type}{The type of the grid}

\item{buffer}{Boolean if the grid should be bigger than the area}
}
\value{
grid from settings
}
\description{
Get grid from settings
}
