% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mod.R
\name{getGridFromShp}
\alias{getGridFromShp}
\title{Get grid from shape}
\usage{
getGridFromShp(shapefiles)
}
\arguments{
\item{shapefiles}{A shape file set}
}
\value{
grids from shape
}
\description{
Get grid from shape
}
