% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mod.R
\name{grid_mod_server}
\alias{grid_mod_server}
\title{Grid module server}
\usage{
grid_mod_server(id, pbd, polygonDraw)
}
\arguments{
\item{id}{Same as id for grid_mod_ui(id)}

\item{pbd}{A reactiveValue}

\item{polygonDraw}{A reactiveValue}
}
\value{
A reactiveValue with the structure \code{reactiveValues(layers = list(others = list(), grids = list()))}
  Each object in the other and grid lists are named with the name they should have in the application and the the list should hold the \code{SpatialPolygons}.
  \code{others} are objects, such as study area, observations...
  \code{grids} are grids.
}
\description{
Create the server side module for creating grids
}
\details{
\code{pbd} Should have have the strucure as: \code{reactiveValues(data=NULL, organised=NULL, visits = NULL, summary = NULL, exportDef = NULL, export = NULL)}, where the variables \code{organised} is important and should be a \code{OrganizedBirds-class}.
\code{polygonDraw} Should have the structure as: \code{reactiveValues(polygon = NULL)}, where \code{polygon} is a \code{SpatialPolygons}. Generated from \code{map_mod_server()}.
}
