% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mod.R
\name{grid_mod_ui}
\alias{grid_mod_ui}
\title{Grid handling UI}
\usage{
grid_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{

}
\description{
Grid handling UI
}
