% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{htmlEscape}
\alias{htmlEscape}
\title{Escape HTML entities}
\source{
htmlTools v 0.5.1.1
}
\usage{
htmlEscape(text, attribute = FALSE)
}
\arguments{
\item{text}{Text to escape}

\item{attribute}{Escape for use as an attribute value}
}
\value{
Character vector with escaped text.
}
\description{
Escape HTML entities contained in a character vector so that it can be safely
included as text or an attribute value within an HTML document
Copied from htmlTools v 0.5.1.1
}
