% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mod.R
\name{layer_mod_server}
\alias{layer_mod_server}
\title{Layer handling server}
\usage{
layer_mod_server(id, layers)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{layers}{Reactive value with the layers to show}
}
\value{
layers outputs
}
\description{
Layer handling server
}
