% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mod.R
\name{layer_mod_ui}
\alias{layer_mod_ui}
\title{Layer handling UI}
\usage{
layer_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{
layers UI
}
\description{
Layer handling UI
}
