% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_page.R
\name{map_page_server}
\alias{map_page_server}
\title{Map page module server}
\usage{
map_page_server(id, pbd_data)
}
\arguments{
\item{id}{Same as id for grid_mod_ui(id)}

\item{pbd_data}{A reactiveValue}
}
\value{
A reactiveValue with the structure \code{reactiveValues(layers = list(others = list(), grids = list()))}
  Each object in the other and grid lists are named with the name they should have in the application and the the list should hold the \code{SpatialPolygons}.
  \code{others} are objects, such as study area, observations...
  \code{grids} are grids.
}
\description{
Creates the map page
}
\details{
\code{pbd} Should have have the strucure as: \code{reactiveValues(data=NULL, organised=NULL, visits = NULL, summary = NULL, exportDef = NULL, export = NULL)}, where the variables \code{organised} is important and should be a \code{OrganizedBirds-class}.
\code{visits} Should be from \code{exploreVisits()}
}
