% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resView.R
\name{resView_mod_server}
\alias{resView_mod_server}
\title{RESULTS server}
\usage{
resView_mod_server(id, toView)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{toView}{Reactive value with the layers to show}
}
\value{

}
\description{
RESULTS server
}
