% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resView.R
\name{resView_mod_ui}
\alias{resView_mod_ui}
\title{Result view UI}
\usage{
resView_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{
results view ui
}
\description{
Result view UI
}
