% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mod.R
\name{summary_mod_server}
\alias{summary_mod_server}
\title{Summary server}
\usage{
summary_mod_server(id, pbd, layersFromMap)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{pbd}{A reactive value with primary biodiversity data}

\item{layersFromMap}{Reactive value with the layers from the map}
}
\value{
summary outputs
}
\description{
Summary server
}
