% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mod.R
\name{summary_mod_ui}
\alias{summary_mod_ui}
\title{Summary UI}
\usage{
summary_mod_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{
summary ui
}
\description{
Summary UI
}
