% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_page.R
\name{summary_page_server}
\alias{summary_page_server}
\title{Summary page server}
\usage{
summary_page_server(id, pbd, layers)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{pbd}{A reactive value with primary biodiversity data}

\item{layers}{A reactive value with the layers to show.}
}
\value{

}
\description{
Summary page server
}
