% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_page.R
\name{summary_page_ui}
\alias{summary_page_ui}
\title{Create th summary page UI}
\usage{
summary_page_ui(id)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}
}
\value{

}
\description{
Create th summary page UI
}
