% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{TimeResCode}
\alias{TimeResCode}
\title{Time resolution codes}
\usage{
TimeResCode()
}
\value{
A vector with \code{c("", "yearly", "month","monthly", "daily")}
}
\description{
Time resolution codes
}
