% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{tooltipHTML}
\alias{tooltipHTML}
\title{Create tooltip}
\usage{
tooltipHTML(text, tip)
}
\arguments{
\item{text}{Text to show}

\item{tip}{Tip to show}
}
\value{
an HTML string for class bubble
}
\description{
Create tooltip
}
