% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_mod.R
\name{code_mod_server}
\alias{code_mod_server}
\title{Code server}
\usage{
code_mod_server(id, inputArg, orgVars, visDat, remVars, layers, summary)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{orgVars}{A reactive value with parameters to organise the data}

\item{visDat}{A reactive value with data on the visits}

\item{remVars}{A reactive value with parameters to remove observations}

\item{layers}{A reactive value with names of the layers to be used}

\item{summary}{A reactive value with parameters to summaryse}

\item{inputArgs}{Reactive value with the input values}
}
\value{

}
\description{
Code server
}
