% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mod.R
\name{map_mod_server}
\alias{map_mod_server}
\title{Export parameters server}
\usage{
map_mod_server(id, layers, pbd_data)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{layers}{Reactive value with the layers to show}

\item{pbd_data}{A reactive value with primary biodiversity data}
}
\value{

}
\description{
Export parameters server
}
