% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expParam_mod.R
\name{obsIndexUI}
\alias{obsIndexUI}
\title{Observation index UI}
\usage{
obsIndexUI(session, spList)
}
\arguments{
\item{session}{The server session}

\item{spList}{A species list}
}
\value{

}
\description{
Observation index UI
}
