% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_page.R
\name{summary_page_server}
\alias{summary_page_server}
\title{Summary page server}
\usage{
summary_page_server(id, pbd, layers, inputArg, orgVars, visDat, remVars)
}
\arguments{
\item{id}{The \code{input} that refers to the UI.}

\item{pbd}{A reactive value with primary biodiversity data}

\item{layers}{A reactive value with names of the layers to be used}

\item{inputArg}{Reactive value with the input values}

\item{orgVars}{A reactive value with parameters to organise the data}

\item{visDat}{A reactive value with data on the visits}

\item{remVars}{A reactive value with parameters to remove observations}
}
\description{
Summary page server
}
